Remove unused SCons and e2 artefacts, move tests
They will be accessible prior to this commit, if needed.
