+++ tests/integration/usbmidi/test_midi_config.py
