+++ tests/integration/usbmidi/test_usb_config.py
