+++ tests/unit/mocks/mock_encoder.cpp
